+++ Js2IL.Tests/Function/JavaScript/Function_Call_Spread_EvaluationOrder.js
@@ -1,0 +1,20 @@
+"use strict";
+
+function dump() {
+    console.log(arguments.length);
+    for (let i = 0; i < arguments.length; i++) {
+        console.log(arguments[i]);
+    }
+}
+
+function arg(x) {
+    console.log("eval" + x);
+    return x;
+}
+
+function spread() {
+    console.log("spread");
+    return [2, 3];
+}
+
+dump(arg(1), ...spread(), arg(4));